// confckurator/src/api/model/Pack.ts
/**
 * Confckurator
 * Confckurator Client/Server API
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface Pack {
    id: number;

    name: string;

  instances: Array<models.Instance>;

}