// confckurator/src/api/model/Instance.ts
/**
 * Confckurator
 * Confckurator Client/Server API
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface Instance {
  /**
   * Persistent instance ID.
   */
  id: number;

  /**
   * Name of the template instance.
   */
  name: string;

  /**
   * Used values.
   */
  values: Array<models.Value>;

  /**
   * Persistent template ID.
   */
  template_id: number;

  /**
   * Name of the used template.
   */
  template_name: string;

}